% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasoline.R
\docType{data}
\name{gasoline}
\alias{gasoline}
\title{NIR spectra of gasoline}
\usage{
data(gasoline)
}
\description{
NIR spectra of gasoline. 60 samples with 401 wavelengths in the interval 900nm-1700nm.
}
\references{
Kalivas, John H. "Two data sets of near infrared spectra." Chemometrics and Intelligent Laboratory Systems 37.2 (1997): 255-259.
}
