% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRutilities.R
\name{plot.TRSVDModel}
\alias{plot.TRSVDModel}
\alias{summary.TRSVDModel}
\title{Specialized X-Y Plotting}
\usage{
\method{plot}{TRSVDModel}(x, which = c(1, 2, 3, 4),
  caption = c("Raw data", "Accuracy", "Regression coef.",
  "Reference vs predicted"), xvals = NULL, log = "x", main = "",
  response = 1, ask = prod(par("mfcol")) < length(which) &&
  dev.interactive(), ...)
}
\arguments{
\item{x}{TRSVDModel object (extracted from x if y is missing).}

\item{which}{choice of response coefficients if more than one.}

\item{xvals}{the coordinates of lines in the plot (\code{xlabs} from y if missing).}

\item{log}{logarithmic plot, defaults to "x".}

\item{...}{arguments to be passed to \code{plot} or \code{lines}.}

\item{y}{not used.}

\item{what}{plot type, defaults to 'prediction', alternatively 'coefficients'.}

\item{object}{TRSVDModel object.}
}
\value{

}
\description{
Specialized X-Y Plotting
}
\examples{
data("gasoline")
X       <- as.matrix(gasoline[, 2])
Y       <- as.matrix(gasoline[, 1])
lambdas <- 10^seq(-2,5,len=1000)
TRModel <- TRReg(X, Y, dtype=1, lambdas)
summary(TRModel)
old.par <- par(mfrow = c(2,2), mar = c(3.5,3.5,2.25,3.25), mgp = c(2.25, 1, 0))
plot(TRModel)
par(old.par)

}
