% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRbcoefs.R
\name{TRbcoefs}
\alias{TRbcoefs}
\title{TRbcoefs}
\usage{
TRbcoefs(decomp, lambdas, Y, centerResponse = TRUE)
}
\arguments{
\item{decomp}{TRSVD object created from TRSVDDecomp.}

\item{lambdas}{Vector of candidate regularization parameter values.}

\item{Y}{response(s).}

\item{centerResponse}{Boolean. Center response(s)? Defaults to TRUE.}
}
\value{
Returns an array of size (p+1)*nlambdas*g containing the regression coefficients.
The first element of the first dimension is the constant term.
}
\description{
Calculates regression coefficients solving a regularized least squares problem.
}
\details{
The function calculates regression coefficients for the regularized least squares problem(s) given by the
TRSVD object, the response(s) and the regularization paramter value(s).
}
\examples{
data("gasoline")
X       <- as.matrix(gasoline[, 2])
Y       <- as.matrix(gasoline[, 1])
decomp  <- TRSVDDecomp(X,dtype=1)
bcoef   <- TRbcoefs(decomp, 10, Y)

}
\seealso{
\code{\link{TRReg}}, \code{\link{TRSVDDecomp}}, \code{\link{TRkFoldCV}}, \code{\link{TRLmatrix}},
\code{\link{TR1SE}}, \code{\link{TRchi2Test}}, \code{\link{TRSVCV}}, \code{\link{TRNumOpt}},
\code{\link{TRPlegendre}}, \code{\link{TRLeverage}}, \code{\link{TRYhatCV}}, \code{\link{TRbcoefs}},
\code{\link{TRPRESS}}, \code{\link{TRGCV}}, \code{\link{TRPred}}, \code{\link{TRDiff}}
}
