% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRchi2Test.R
\name{TRchi2Test}
\alias{TRchi2Test}
\title{TRchi2Test}
\usage{
TRchi2Test(decomp, lambdas, rmsecv, Y, alpha = 0.2)
}
\arguments{
\item{decomp}{Output from TRSVDDecomp}

\item{lambdas}{Vector of candidate regularization parameter values}

\item{rmsecv}{RMSECV values corresponding to the lambda vector}

\item{Y}{Response matrix/vector}

\item{alpha}{Significance level. Defaults to 0.2.}
}
\value{
rmsecvoptchi2 3*g matrix containing for each response the selected regularization parameter value, the index of this regularization
parameter value, as well as the associated RMSECV value.
}
\description{
Uses a Chi^2-test to select a simpler model than the PRESS-minimal model for regularized least squares problems.
}
\details{
A Chi^2-test is used to select a simpler model than the PRESS-minimal model for regularized least squares problems
by selecting the model with the largest regularization parameter value that is not significantly different
(as determined by the significance level) from the PRESS-optimal model.
}
\examples{
data("gasoline")
X          <- as.matrix(gasoline[, 2])
Y          <- as.matrix(gasoline[, 1])
lambdas    <- 10^seq(-2,5,len=1000)
TRModel    <- TRReg(X, Y, dtype=1, lambdas)
chi2Output <- TRchi2Test(TRModel$decomp, lambdas, TRModel$rmsecv, Y, alpha=0.2)
bchi2      <- TRbcoefs(TRModel$decomp, chi2Output[1], Y)

}
\seealso{
\code{\link{TRReg}}, \code{\link{TRSVDDecomp}}, \code{\link{TRkFoldCV}}, \code{\link{TRLmatrix}},
\code{\link{TR1SE}}, \code{\link{TRchi2Test}}, \code{\link{TRSVCV}}, \code{\link{TRNumOpt}},
\code{\link{TRPlegendre}}, \code{\link{TRLeverage}}, \code{\link{TRYhatCV}}, \code{\link{TRbcoefs}},
\code{\link{TRPRESS}}, \code{\link{TRGCV}}, \code{\link{TRPred}}, \code{\link{TRDiff}}
}
