% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRDiff.R
\name{TRDiff}
\alias{TRDiff}
\title{TRDiff}
\usage{
TRDiff(p, dtype)
}
\arguments{
\item{p}{Size of matrix}

\item{dtype}{Derivative order}
}
\value{
sparse Matrix of discrete derivatives.
}
\description{
Function for creating square discrete derivative matrix of given size and degree.
}
\details{
The function is used to create square discrete derivative matrices to be used as regularization matrices in regularized
least squares problems.
The function is similar to "diff" in MATLAB.
}
\examples{
L <- TRDiff(401, 1)

}
\seealso{
\code{\link{TRReg}}, \code{\link{TRSVDDecomp}}, \code{\link{TRkFoldCV}}, \code{\link{TRLmatrix}},
\code{\link{TR1SE}}, \code{\link{TRchi2Test}}, \code{\link{TRSVCV}}, \code{\link{TRNumOpt}},
\code{\link{TRPlegendre}}, \code{\link{TRLeverage}}, \code{\link{TRYhatCV}}, \code{\link{TRbcoefs}},
\code{\link{TRPRESS}}, \code{\link{TRGCV}}, \code{\link{TRPred}}, \code{\link{TRDiff}}
}
