% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRkFoldCV.R
\name{TRkFoldCV}
\alias{TRkFoldCV}
\title{TR k-fold CV}
\usage{
TRkFoldCV(X, Y, dtype = 0, lambdas = 10^seq(-5, 20, len = 1000),
  nFolds = 5)
}
\arguments{
\item{X}{data frame (predictors)}

\item{Y}{data frame (response(s))}

\item{dtype}{Regularization type. -1, 0, or positive integer. -1=Standardization, 0=L_2 regularization,
and positive integer gives a discrete derivative matrix of order dtype.}

\item{lambdas}{Vector of candidate regularization parameter values.}

\item{nFolds}{Positive integer. Number of folds to be used in cross-validation or
and indexing vector for the folds. Defaults to 5.}
}
\value{
Returns a list consisting of (i) The regression coefficients minimising RMSECV,
(ii) RMSECV-values for all the regularization parameter values given as input,
(iii) a 3*g matrix of optimal models (for each response the
press-minimal lambda-value is given, the index of the press-minimal lambda value, as well as the associated RMSECV value),
(iv) TRSVD object associated with the full X-matrix
}
\description{
Does k-fold cross-validation for regularized least squares problems.
}
\details{
The function does k-fold cross-validation for regularized least squares problems.
The desired regularization matrix is indicated as input together with the candidate regularization parameter values.
Multivariate problems are treated as a series of univariate problems.
The function finds PRESS-minimal and GCV-minimal regression coefficients.
}
\examples{
data("gasoline")
X       <- as.matrix(gasoline[, 2])
Y       <- as.matrix(gasoline[, 1])
lambdas <- 10^seq(-2,5,len=1000)
TRModel <- TRkFoldCV(X, Y, dtype=1, lambdas, nFolds=5)

}
\seealso{
\code{\link{TRReg}}, \code{\link{TRSVDDecomp}}, \code{\link{TRkFoldCV}}, \code{\link{TRLmatrix}}, 
\code{\link{TR1SE}}, \code{\link{TRchi2Test}}, \code{\link{TRSVCV}}, \code{\link{TRNumOpt}},
\code{\link{TRPlegendre}}, \code{\link{TRLeverage}}, \code{\link{TRYhatCV}}, \code{\link{TRbcoefs}},
\code{\link{TRPRESS}}, \code{\link{TRGCV}}, \code{\link{TRPred}}, \code{\link{TRDiff}}
}
