% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRLeverage.R
\name{TRLeverage}
\alias{TRLeverage}
\title{TRLeverage}
\usage{
TRLeverage(decomp, lambdas)
}
\arguments{
\item{decomp}{TRSVD object created from TRSVDDEcomp.}

\item{lambdas}{Vector of regularization parameter values}
}
\value{
H n*nlambdas matrix of leverage values.
}
\description{
Calculates leverage values given a SVDDecomp object and a vector of regularization parameter values.
}
\details{
The function uses calculates the leverage values (diagonal elements of the projection matrix) for the
regularised least squares problem associated with the given SVDDecomp object and the given vector of
regularization parameter values (lambdas). These leverage values can be used to efficiently find
the PRESS-statistic and GCV using rank 1 updates.
}
\examples{
data("gasoline")
X       <- as.matrix(gasoline[, 2])
decomp  <- TRSVDDecomp(X,dtype=1)
lambdas <- 10^seq(-2,5,len=1000)
H       <- TRLeverage(decomp, lambdas)

}
\seealso{
\code{\link{TRReg}}, \code{\link{TRSVDDecomp}}, \code{\link{TRkFoldCV}}, \code{\link{TRLmatrix}}, 
\code{\link{TR1SE}}, \code{\link{TRchi2Test}}, \code{\link{TRSVCV}}, \code{\link{TRNumOpt}},
\code{\link{TRPlegendre}}, \code{\link{TRLeverage}}, \code{\link{TRYhatCV}}, \code{\link{TRbcoefs}},
\code{\link{TRPRESS}}, \code{\link{TRGCV}}, \code{\link{TRPred}}, \code{\link{TRDiff}}
}
