% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRLmatrix.R
\name{TRLmatrix}
\alias{TRLmatrix}
\title{TRLmatrix}
\usage{
TRLmatrix(X, dtype = 0, epsilon = 1e-10, appendLegendre = TRUE)
}
\arguments{
\item{X}{Data matrix or number of columns in data matrix (the matrix itself is necessary when dtype=-1)}

\item{dtype}{Positive integer. -1 gives standardization matrix, 0 gives empty matrix (for L_2 regularization), and a
positive integer gives a discrete derivative matrix of order dtype.}

\item{epsilon}{If dtype>=1 the extra rows appended to obtain full rank will be scaled by sqrt(epsilon). Defaults to epsilon=1e-10.}

\item{appendLegendre}{Boolean. Append Legendre polynomials to L matrix to obtain a full rank regularization matrix? Defaults to TRUE.}
}
\value{
Regularization matrix for TR problem
}
\description{
Function to construct common regularization matrices used in regularized least squares problems.
}
\details{
The function can construct common regularization matrices used in regularized least squares problems.
Standardization as well as discrete derivative operators are supported.
Discrete derivatives are augmented with Legendre polynomials to obtain a full rank matrix. The Legendre
polynomials are scaled according to the square root of the epsilon argument.
For L_2 regularization an empty data frame is returned.
}
\examples{
L <- TRLmatrix(401, 1, 1e-10)

}
\seealso{
\code{\link{TRReg}}, \code{\link{TRSVDDecomp}}, \code{\link{TRkFoldCV}}, \code{\link{TRLmatrix}},
\code{\link{TR1SE}}, \code{\link{TRchi2Test}}, \code{\link{TRSVCV}}, \code{\link{TRNumOpt}},
\code{\link{TRPlegendre}}, \code{\link{TRLeverage}}, \code{\link{TRYhatCV}}, \code{\link{TRbcoefs}},
\code{\link{TRPRESS}}, \code{\link{TRGCV}}, \code{\link{TRPred}}, \code{\link{TRDiff}}
}
