% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRNumOpt.R
\name{TRNumOpt}
\alias{TRNumOpt}
\title{TR Numerical Optimisation}
\usage{
TRNumOpt(decomp, y, criterion = "PRESS", centerResponse = TRUE,
  minbnd = 0, maxbnd = 1e+20)
}
\arguments{
\item{decomp}{TRSVD object created from TRSVDdecomp}

\item{y}{Response vector}

\item{criterion}{PRESS or GCV. Minimise PRESS or GCV? Defaults to PRESS}

\item{centerResponse}{Center the response variable? Defaults to TRUE.}

\item{minbnd}{Lower bound in fminbnd. Defaults to zero.}

\item{maxParam}{Upper bound in fminbnd. Defaults to 1e20.}
}
\value{
optLambda Output from fminbnd.
}
\description{
Function uses fminbnd to find the regularization parameter value minimising PRESS or GCV for regularized least squares problems.
The function only supports univariate responses.
}
\examples{
data("gasoline")
X           <- as.matrix(gasoline[, 2])
Y           <- as.matrix(gasoline[, 1])
decomp      <- TRSVDDecomp(X,dtype=1)
optLambda   <- TRNumOpt(decomp, Y, "PRESS")

}
\seealso{
\code{\link{TRReg}}, \code{\link{TRSVDDecomp}}, \code{\link{TRkFoldCV}}, \code{\link{TRLmatrix}}, 
\code{\link{TR1SE}}, \code{\link{TRchi2Test}}, \code{\link{TRSVCV}}, \code{\link{TRNumOpt}},
\code{\link{TRPlegendre}}, \code{\link{TRLeverage}}, \code{\link{TRYhatCV}}, \code{\link{TRbcoefs}},
\code{\link{TRPRESS}}, \code{\link{TRGCV}}, \code{\link{TRPred}}, \code{\link{TRDiff}}
}
