% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRPlegendre.R
\name{TRPlegendre}
\alias{TRPlegendre}
\title{TR Plegendre}
\usage{
TRPlegendre(d, p)
}
\arguments{
\item{d}{non-negative integer. Degree of polynomials}

\item{p}{positive integer. Size of vectors}
}
\value{
matrix containing the orthonormal Legendre polynomials.
}
\description{
Creates Legendre polynomials of given degree and size using a QR factorisation.
}
\details{
The function creates orthonormal Legendre polynomials of length p and degree d.
This is done by sampling the function x^i uniformly at p points in the interval
[-1,1]. Orthogonal vectors are then obtained by a QR factorisation.
}
\examples{
Polynomials <- TRPlegendre(2,401)

}
\seealso{
\code{\link{TRReg}}, \code{\link{TRSVDDecomp}}, \code{\link{TRkFoldCV}}, \code{\link{TRLmatrix}},
\code{\link{TR1SE}}, \code{\link{TRchi2Test}}, \code{\link{TRSVCV}}, \code{\link{TRNumOpt}},
\code{\link{TRPlegendre}}, \code{\link{TRLeverage}}, \code{\link{TRYhatCV}}, \code{\link{TRbcoefs}},
\code{\link{TRPRESS}}, \code{\link{TRGCV}}, \code{\link{TRPred}}, \code{\link{TRDiff}}
}
