% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRPred.R
\name{TRPred}
\alias{TRPred}
\title{TR Prediction}
\usage{
TRPred(X, bcoefs, Y = FALSE)
}
\arguments{
\item{X}{data frame (predictors).}

\item{bcoefs}{Regression coefficients.}

\item{Y}{(Optional) True response(s). Defaults to FALSE}
}
\value{
If the true response is not given as input the function returns a list consisting
of Yhat calculated from the given data and regression coefficients. If the true response is
given as input then the output list also contains the RMSEP and squared prediction error.
}
\description{
Function for prediction given predictors and regression coefficients. If the true response
is given as input then the prediction error will also be calculated.
}
\examples{
data("gasoline")
X       <- as.matrix(gasoline[, 2])
Y       <- as.matrix(gasoline[, 1])
lambdas <- 10^seq(-2,5,len=1000)
TRModel <- TRReg(X, Y, dtype=1, lambdas)
TRPred  <- TRPred(X, TRModel$b, Y)

}
\seealso{
\code{\link{TRReg}}, \code{\link{TRSVDDecomp}}, \code{\link{TRkFoldCV}}, \code{\link{TRLmatrix}}, 
\code{\link{TR1SE}}, \code{\link{TRchi2Test}}, \code{\link{TRSVCV}}, \code{\link{TRNumOpt}},
\code{\link{TRPlegendre}}, \code{\link{TRLeverage}}, \code{\link{TRYhatCV}}, \code{\link{TRbcoefs}},
\code{\link{TRPRESS}}, \code{\link{TRGCV}}, \code{\link{TRPred}}, \code{\link{TRDiff}}
}
