% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRRes.R
\name{TRRes}
\alias{TRRes}
\title{TRRes}
\usage{
TRRes(decomp, lambdas, Y, centerResponse = TRUE)
}
\arguments{
\item{decomp}{TRSVD object created from TRSVDDecomp}

\item{lambdas}{Vector of candidate regularization parameter values}

\item{Y}{response vector}

\item{centerResponse}{Boolean. Center response(s)? Defaults to TRUE.}
}
\value{
The output is a a list with an array of dimensions (n,nlambdas,g) (number of samples, regularization parameter values, and responses),
and a n*nlambdas matrix of denominator factors used by many of the functions in the package.
}
\description{
Calculates the residuals for regularized least squares problem.
}
\details{
Given a TRSVD object, a vector of regularization parameter values, and a response matrix/vector the
function calculates the residuals of the regularized least squares problem implied by  the TRSVD object
for the regularization parameter values contained in the vector lambdas.
}
\examples{
data(gasoline)
X           <- as.matrix(gasoline[, 2])
Y           <- as.matrix(gasoline[, 1])
decomp      <- TRSVDDecomp(X,dtype=1)
lambdas     <- 10^seq(-2,5,len=1000)
TRResOutput <- TRRes(decomp, lambdas, Y)

}
\seealso{
\code{\link{TRReg}}, \code{\link{TRSVDDecomp}}, \code{\link{TRkFoldCV}}, \code{\link{TRLmatrix}},
\code{\link{TR1SE}}, \code{\link{TRchi2Test}}, \code{\link{TRSVCV}}, \code{\link{TRNumOpt}},
\code{\link{TRPlegendre}}, \code{\link{TRLeverage}}, \code{\link{TRYhatCV}}, \code{\link{TRbcoefs}},
\code{\link{TRPRESS}}, \code{\link{TRGCV}}, \code{\link{TRPred}}, \code{\link{TRDiff}}
}
