% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRSVDDecomp.R
\name{TRSVDDecomp}
\alias{TRSVDDecomp}
\title{TR SVD Decomposition}
\usage{
TRSVDDecomp(X, dtype = 0, addCrit = FALSE, center = TRUE)
}
\arguments{
\item{X}{Data matrix.}

\item{dtype}{-1, 0, or positive integer. Controls regularization matrix. -1=Standardization, 0=L_2 regularization,
and positive integer gives a discrete derivative matrix of order dtype.}

\item{addCrit}{Any additional criteria that should be appended to the X matrix. Defaults to FALSE.}

\item{center}{Boolean. Center data matrix prior to finding the SVD? Defaults to TRUE}
}
\value{
decomp TRSVD object. Contains the SVD, the regularization matrix, and more.
}
\description{
Find SVD of a data matrix to allow for efficient computation of PRESS, GCV, etc for a regularized least squares problem.
}
\details{
The function finds the singular value decomposition (SVD) of the given data matrix. This decomposition
can be used with the other TR functions to efficiently calculate PRESS, GCV, and so on for regularized least squares problems.
Additional criteria on the regression coefficients are appended to the data matrix prior to finding the SVD.
}
\examples{
data("gasoline")
X      <- as.matrix(gasoline[, 2])
decomp <- TRSVDDecomp(X,dtype=1)

}
\seealso{
\code{\link{TRReg}}, \code{\link{TRSVDDecomp}}, \code{\link{TRkFoldCV}}, \code{\link{TRLmatrix}},
\code{\link{TR1SE}}, \code{\link{TRchi2Test}}, \code{\link{TRSVCV}}, \code{\link{TRNumOpt}},
\code{\link{TRPlegendre}}, \code{\link{TRLeverage}}, \code{\link{TRYhatCV}}, \code{\link{TRbcoefs}},
\code{\link{TRPRESS}}, \code{\link{TRGCV}}, \code{\link{TRPred}}, \code{\link{TRDiff}}
}
