% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRYhatCV.R
\name{TRYhatCV}
\alias{TRYhatCV}
\title{TRYhatCV}
\usage{
TRYhatCV(decomp, lambdas, Y, H = TRLeverage(decomp, lambdas))
}
\arguments{
\item{decomp}{TRSVD object created from TRSVDDecomp}

\item{lambdas}{Vector of candidate regularization parameter values}

\item{Y}{response vector}

\item{H}{n*nlambdas matrix of leverage values. Calculated on the fly if not given as input.}
}
\value{
Returns a list consisting of the fitted responses, the cross-validated responses,
the fitted residuals, and the cross-validated residuals.
}
\description{
Function for calculating training set estimates/residuals and Loo estimate/residuals associated
with a regularized least squares problem.
}
\details{
Given a TRSVD object, response(s) and regularization parameter value(s), the function finds
the fitted Y-values, the LooCV Y-values, the fitted residuals, and the LooCV cross-validated residuals
solving the least squares problem implied by the input.
}
\examples{
data("gasoline")
X        <- as.matrix(gasoline[, 2])
Y        <- as.matrix(gasoline[, 1])
decomp   <- TRSVDDecomp(X,dtype=1)
lambdas  <- 10^seq(-2,5,len=1000)
CVOutput <- TRYhatCV(decomp, lambdas, Y)

}
\seealso{
\code{\link{TRReg}}, \code{\link{TRSVDDecomp}}, \code{\link{TRkFoldCV}}, \code{\link{TRLmatrix}}, 
\code{\link{TR1SE}}, \code{\link{TRchi2Test}}, \code{\link{TRSVCV}}, \code{\link{TRNumOpt}},
\code{\link{TRPlegendre}}, \code{\link{TRLeverage}}, \code{\link{TRYhatCV}}, \code{\link{TRbcoefs}},
\code{\link{TRPRESS}}, \code{\link{TRGCV}}, \code{\link{TRPred}}, \code{\link{TRDiff}}
}
